requirements.txt
discord.py==1.2.1
NyaaPy
gspread
oauth2client
